# perf: cache results for GHSA queries